Functions

Q-Voter Model

QVoterSeeding.qvoter — Function

qvoter(g::SimpleGraph{<:Integer}, active::BitVector, q::Integer; ε::AbstractFloat = 0.0, flexible::BitVector = trues(nv(g)), replace::Bool = false, warm_up::Integer = 10^5, averaging_period::Integer = 10^3)

Run a Monte Carlo simulation of the q-voter model.

Arguments

g: Graph representing the structure of the agent system.

active: BitVector representing states of the agents (the vector is mutated during simulation).

q: Number of neighbors in the influence group.

Keyword arguments

ε = 0.0: Probability of flipping state when q-panel is not unanimous.

flexible = trues(nv(g)): BitVector represeting flexibility of the agents, i.e., whether an agent changes its state during simulation.

replace = false: If true, cosntruct the q-panel by drawing neighbors with replacement, otherwise draw without replacement.

warm_up = 10^5: Number of Monte Carlo steps calculated during the simulation warm up.

averaging_period = 10^3: Number of Monte Carlo steps over which the concentration is averaged.

Return the final concentration of active agents. If the system reaches an absorbing state, the final concentration is returned immediately.

Seeding

QVoterSeeding.seeding — Function

seeding(network::Union{Symbol, String}, strategy::Symbol, zealots::Bool, q::Integer, budgets::AbstractVector{<:AbstractFloat}, rng::AbstractRNG = Random.default_rng())

Run network seeding experiments with q-voter dynamics with for a specified seeding strategy and seeding budgets. If zealots = true, seeds will act as zealots.

Arguments network and rng are passed to loadnetwork.

Available options for strategy:

:hd for high degree

:pr for PageRank

:cc for complex centrality

:onehop for one-hop

:random for random.

Return a vector of final concentrations of active vertices for each seeding budget.

Networks

QVoterSeeding.loadnetwork — Function

loadnetwork(networkname::Symbol, rng::AbstractRNG = Random.default_rng(); preprocess::Bool = true)

Load network data from the networks directory or from the specified filepath.

Available options for network:

:snap for SNAP Facebook network

:politicians for GEMSEC Facebook page network of politicians

"..." for network at filepath "..."; the file is assumed to be an edge list in a CSV format with node indexing starting from 0.

Optional argument rng specifies a random number generator used for preprocessing of the network data. Keyword argument preprocess specifies whether to perform preprocessing (prune! and addneighbors!).

Return a graph representing the selected network.

QVoterSeeding.average_degree — Function

average_degree(network::Union{Symbol, String}, m::Integer, rng::AbstractRNG = Random.default_rng())

Calculate the average node degree of a raw network and an ensemble of m preprocessed networks.

Available options for network:

:snap for SNAP Facebook network

:politicians for GEMSEC Facebook page network of politicians

"..." for network file at path "...".

Optional argument rng specifies a random number generator used for generating an ensemble of preprocessed networks.

Return a tuple of the average node degree for a raw network and for a network ensemble.

QVoterSeeding.average_clustering — Function

average_clustering(network::Union{Symbol, String}, m::Integer, rng::AbstractRNG = Random.default_rng())

Calculate the average clustering coefficient of a raw network and an ensemble of m preprocessed networks.

Available options for network:

:snap for SNAP Facebook network

:politicians for GEMSEC Facebook page network of politicians

"..." for network file at path "...".

Optional argument rng specifies a random number generator used for generating an ensemble of preprocessed networks.

Return a tuple of the average clustering coefficient for a raw network and for a network ensemble.

QVoterSeeding.average_pathlen — Function

average_pathlen(network::Union{Symbol, String}, m::Integer, rng::AbstractRNG = Random.default_rng())

Calculate the average shortest path length of a raw network and an ensemble of m preprocessed networks.

Available options for network:

:snap for SNAP Facebook network

:politicians for GEMSEC Facebook page network of politicians

"..." for network file at path "...".

Optional argument rng specifies a random number generator used for generating an ensemble of preprocessed networks.

Return a tuple of the average shortest path length for a raw network and for a network ensemble.

QVoterSeeding.prune! — Function

prune!(g::SimpleGraph{<:Integer})

Modify graph g by removing self-loops and isolated vertices.

QVoterSeeding.addneighbors! — Function

addneighbors!(g::SimpleGraph{<:Integer}, rng::AbstractRNG = Random.default_rng())

Modify graph g by adding a new neighbor to every single-neighbor-vertex via triad formation.

Optional argument rng specifies a random number generator.

Complex Centrality

QVoterSeeding.complexcentrality — Function

complexcentrality(g::SimpleGraph{<:Integer}, T::Vector{<:Integer}[, v::Integer])

Calculate the complex centrality for graph g with threshold values specifified by vector T.

Return a vector of complex centralities calculated for each vertex. If v is specified, calculate and return only the complex centrality of vertex v.

QVoterSeeding.competitive_complexcentrality — Function

competitive_complexcentrality(g::SimpleGraph{<:Integer}, q::Integer[, v::Integer])

Calculate the complex centrality for graph g with threshold values adjusted to q-voter dynamics, according to: $T_w = \max\{q, \lceil k_w/2\rceil\}$, where $k_w$ is the degree of vertex $w$.

Return a vector of complex centralities calculated for each vertex. If v is specified, calculate and return only the complex centrality of vertex v.

One-hop

QVoterSeeding.onehop — Function

onehop(g::SimpleGraph{<:Integer}, seedsize::Integer)

Construct a seed set of size seedsize on graph g using one-hop strategy.

Return a vector of seed vertices.